+++ tests/phpunit/Unit/LibraryFactoryTest.php
@@ -2,6 +2,10 @@
 
 namespace SMW\Scribunto\Tests\Unit;
 
+use MediaWiki\Language\Language;
+use MediaWiki\Parser\Parser;
+use MediaWiki\Parser\ParserOutput;
+use MediaWiki\Title\Title;
 use PHPUnit\Framework\TestCase;
 use SMW\ParserFunctions\SetParserFunction;
 use SMW\Query\QueryResult;
@@ -11,7 +15,7 @@ use SMW\Scribunto\LibraryFactory;
  * @covers \SMW\Scribunto\LibraryFactory
  * @group semantic-scribunto
  *
- * @license GNU GPL v2+
+ * @license GPL-2.0-or-later
  * @since 1.0
  *
  * @author mwjames
@@ -22,8 +26,7 @@ class LibraryFactoryTest extends TestCase {
 	private $parser;
 
 	protected function setUp(): void {
-
-		$language = $this->getMockBuilder( '\Language' )
+		$language = $this->getMockBuilder( Language::class )
 			->disableOriginalConstructor()
 			->getMock();
 
@@ -37,30 +40,29 @@ class LibraryFactoryTest extends TestCase {
 
 		$this->store->expects( $this->any() )
 			->method( 'getQueryResult' )
-			->will( $this->returnValue( $queryResult ) );
+			->willReturn( $queryResult );
 
 		$stripState = $this->createMock( \StripState::class );
-		$this->parser = $this->createMock( \Parser::class );
+		$this->parser = $this->createMock( Parser::class );
 
 		$this->parser->expects( $this->any() )
 			->method( 'getStripState' )
-			->will( $this->returnValue( $stripState ) );
+			->willReturn( $stripState );
 
 		$this->parser->expects( $this->any() )
 			->method( 'getTitle' )
-			->will( $this->returnValue( \Title::newFromText( 'Foo' ) ) );
+			->willReturn( Title::newFromText( 'Foo' ) );
 
 		$this->parser->expects( $this->any() )
 			->method( 'getOutput' )
-			->will( $this->returnValue( new \ParserOutput() ) );
+			->willReturn( new ParserOutput() );
 
 		$this->parser->expects( $this->any() )
 			->method( 'getTargetLanguage' )
-			->will( $this->returnValue( $language ) );
+			->willReturn( $language );
 	}
 
 	public function testCanConstruct() {
-
 		$this->assertInstanceOf(
 			LibraryFactory::class,
 			new LibraryFactory( $this->store )
@@ -68,7 +70,6 @@ class LibraryFactoryTest extends TestCase {
 	}
 
 	public function testCanConstructQueryResult() {
-
 		$instance = new LibraryFactory(
 			$this->store
 		);
@@ -80,7 +81,6 @@ class LibraryFactoryTest extends TestCase {
 	}
 
 	public function testCanConstructParserParameterProcessor() {
-
 		$instance = new LibraryFactory(
 			$this->store
 		);
@@ -92,7 +92,6 @@ class LibraryFactoryTest extends TestCase {
 	}
 
 	public function testCanConstructSetParserFunction() {
-
 		$instance = new LibraryFactory(
 			$this->store
 		);
@@ -104,7 +103,6 @@ class LibraryFactoryTest extends TestCase {
 	}
 
 	public function testCanConstructSubobjectParserFunction() {
-
 		$instance = new LibraryFactory(
 			$this->store
 		);
@@ -116,7 +114,6 @@ class LibraryFactoryTest extends TestCase {
 	}
 
 	public function testCanConstructLuaAskResultProcessor() {
-
 		$queryResult = $this->getMockBuilder( QueryResult::class )
 			->disableOriginalConstructor()
 			->getMock();